[#214] Extract "options" of "ReaderSettings" to new component
